+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - connect to the demo server
 - trial period of 30 days for the demo server
 - transfer user region in conversion
+- editable extensions on the settings page
 
 ## 2.4.0
 ## Added